+++ notification-support/src/main/java/com/expediagroup/streamplatform/streamregistry/core/events/NotificationEventHandler.java
@@ -15,10 +15,52 @@
  */
 package com.expediagroup.streamplatform.streamregistry.core.events;
 
+import java.util.Optional;
+import java.util.function.Function;
+
+import lombok.val;
+
+import org.springframework.kafka.support.KafkaHeaders;
+import org.springframework.messaging.Message;
+import org.springframework.messaging.support.MessageBuilder;
+
+import com.expediagroup.streamplatform.streamregistry.core.events.config.NotificationEventConstants;
+
 public interface NotificationEventHandler<T> {
   void onCreate(NotificationEvent<T> event);
 
   void onUpdate(NotificationEvent<T> event);
 
   void onDelete(NotificationEvent<T> event);
+
+  static <T, V> T sendEntityNotificationEvent(
+      Function<V, ?> entityToKeyRecord,
+      Function<V, ?> entityToValueRecord,
+      Function<Message<?>, T> sendMessage,
+      String topic,
+      NotificationEvent<V> event
+  ) {
+    val key = entityToKeyRecord.apply(event.getEntity());
+    val value = entityToValueRecord.apply(event.getEntity());
+
+    val eventType = Optional.ofNullable(event.getEventType())
+        .map(EventType::toString)
+        .orElse(NotificationEventConstants.NOTIFICATION_TYPE_HEADER.defaultValue);
+
+    val entity = Optional.ofNullable(event.getEntity())
+        .map(Object::getClass)
+        .map(Class::getSimpleName)
+        .map(String::toUpperCase)
+        .orElse(NotificationEventConstants.ENTITY_TYPE_HEADER.defaultValue);
+
+    val message = MessageBuilder
+        .withPayload(value)
+        .setHeader(KafkaHeaders.MESSAGE_KEY, key)
+        .setHeader(KafkaHeaders.TOPIC, topic)
+        .setHeader(NotificationEventConstants.NOTIFICATION_TYPE_HEADER.name, eventType)
+        .setHeader(NotificationEventConstants.ENTITY_TYPE_HEADER.name, entity)
+        .build();
+
+    return sendMessage.apply(message);
+  }
 }