+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Emit event on producer mutation.
+- Emit event on producer/consumer mutation.
 - Emit event on streambinding mutation.
 - Added `repository-api` module to define new repository abstraction exposing model instead of data.
 - Added `repository-postgres` module to implement new repository adapting JPA to consolidate JPA/Hibernate implementation details.